moved build api to /api
